requirements.txt
pandas
numpy
regex
beautifulsoup4
python-dateutil
langchain-core
langchain-community
langchain-text-splitters
langchain-openai
